Extract more classes into their own files
