+++ ts/src/relation.ts
@@ -1,8 +1,10 @@
 import { OsmObject } from "./osm-object";
 import { Way } from "./way";
 import { Node } from "./node";
-
-import { first, LateBinder, ptInsidePolygon, RefElements, WayCollection } from "./utils";
+import { WayCollection } from "./way-collection";
+import { LateBinder } from "./late-binder";
+import { RefElements } from "./ref-elements";
+import { first, pointInsidePolygon } from "./utils";
 import type { BBox, Feature, GeometryObject, LineString, MultiLineString, MultiPoint, MultiPolygon, Point, Polygon } from "geojson";
 
 export class Relation extends OsmObject {
@@ -105,7 +107,7 @@ export class Relation extends OsmObject {
 
     public toFeatureArray(): Array<Feature<any, any>> {
         function constructStringGeometry(ws: WayCollection): LineString | MultiLineString | null {
-            const strings = ws ? ws.toStrings() : [];
+            const strings = ws ? ws.mergeWays() : [];
             if (strings.length > 0) {
                 if (strings.length === 1) {
                     return {
@@ -138,7 +140,7 @@ export class Relation extends OsmObject {
                 ring = innerRings.shift();
                 while (ring) {
                     for (const idx in outerRings) {
-                        if (ptInsidePolygon(first(ring), outerRings[idx])) {
+                        if (pointInsidePolygon(first(ring), outerRings[idx])) {
                             compositPolyons[idx].push(ring);
                             break;
                         }
